Working towards 0 flow errors
