+++ client-v2/src/actions/Clipboard.js
@@ -1,5 +1,6 @@
 // @flow
 
+import type { Dispatch } from 'redux';
 import { saveClipboard, getArticles } from 'services/faciaApi';
 import { actions as externalArticleActions } from 'shared/bundles/externalArticlesBundle';
 import { batchActions } from 'redux-batched-actions';
@@ -18,7 +19,7 @@ function updateClipboardContent(clipboardContent: Array<string>) {
 }
 
 function storeClipboardContent(clipboardContent: Array<NestedArticleFragment>) {
-  return (dispatch: Dispatch) => {
+  return (dispatch: Dispatch<any>) => {
     const normalisedClipboard: {
       clipboard: { articles: Array<string> },
       articleFragments: { [string]: ArticleFragment }
