+++ client-v2/src/actions/Fronts.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ThunkAction } from 'types/Store';
+import type { ThunkAction, Dispatch } from 'types/Store';
 import type { Action } from 'types/Action';
 import { batchActions } from 'redux-batched-actions';
 import {
