+++ client-v2/src/components/Clipboard.js
@@ -5,7 +5,6 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import * as Guration from 'lib/guration';
 import type { Edit } from 'lib/guration';
-import { type Dispatch } from 'types/Store';
 import { type State } from 'types/State';
 import { urlToArticle } from 'util/collectionUtils';
 import { clipboardAsTreeSelector } from 'shared/selectors/shared';
