+++ client-v2/src/components/FrontsEdit/CollectionComponents/Supporting.js
@@ -4,6 +4,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import Article from 'shared/components/Article';
 import { removeSupportingArticleFragment } from 'actions/ArticleFragments';
+import type { Dispatch } from 'types/Store';
 
 type SupportingProps = {
   uuid: string,
